+++ python/sglang/srt/models/deepseek_common/utils.py
@@ -1,3 +1,17 @@
+# Copyright 2026 SGLang Team
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+
 from sglang.srt.layers.quantization.fp8_kernel import is_fp8_fnuz
 from sglang.srt.utils import (
     cpu_has_amx_support,
@@ -9,6 +23,8 @@ from sglang.srt.utils import (
     is_hip,
     is_npu,
 )
+from sglang.srt.environ import envs
+from sglang.srt.layers.moe.fused_moe_triton.layer import get_moe_runner_backend
 
 _is_hip = is_hip()
 _is_cuda = is_cuda()
@@ -19,5 +35,33 @@ _is_cpu_amx_available = cpu_has_amx_support()
 _is_cpu = is_cpu()
 _device_sm = get_device_sm()
 _is_gfx95_supported = is_gfx95_supported()
-
 _use_aiter_gfx95 = _use_aiter and _is_gfx95_supported
+
+
+def awq_dequantize_func():
+    '''
+    Get the AWQ dequantize function for the current device
+
+    Return:
+        - The AWQ dequantize function for the current device.
+        - None if the current device is not supported.
+    '''
+    if _is_cuda:
+        from sgl_kernel import awq_dequantize
+        return awq_dequantize
+    elif _is_hip:
+        from sglang.srt.layers.quantization.awq_triton import awq_dequantize_triton as awq_dequantize
+        return awq_dequantize
+    elif _is_npu:
+        from sglang.srt.layers.quantization.awq_triton import awq_dequantize_decomposition as awq_dequantize
+        return awq_dequantize
+    else:
+        return None
+
+def enable_nextn_moe_bf16_cast_to_fp8(quant_config):
+    return (
+        envs.SGLANG_NVFP4_CKPT_FP8_NEXTN_MOE.get()
+        and quant_config is not None
+        and quant_config.get_name() == "modelopt_fp4"
+        and get_moe_runner_backend().is_deep_gemm()
+    )
